Allow only unsealing once
Right now, a manual sealing of the Vault will result in the auto-init
unsealer immediately unsealing it again. This adds a new configuration
option that makes the auto-init unsealer stop after the first success.

This still maintains HA in the event that Kubernetes moves the pod for
rescheduling (since it'll spawn a new container), but solves the problem
where the Vault server is sealed manually by an operator in response to
a security incident.